port to boost::signals2
unfortunately, this needs ugly code duplication:
boost::signals2 has deprecated the old API which we need for compatibility with several (not too old) compilers/platforms, and „hides“ it if and only if the compilers supports the new API (which is, admittedly, nicer).

I tend to think that this is actually highly undesirable behavior (read: a bug), since users caring about both old and new compilers (e.g. VC10+12) need to duplicate the relevant code, because neither API is always available.  At least, there’s a preprocessor define that we can use..